Sécurité likes

+++ back/controllers/sauce.js
@@ -2,6 +2,7 @@ const Sauce = require('../models/sauce');
 
 //Permet de modifier le système de fichiers
 const fs = require('fs');
+const jwt = require('jsonwebtoken');
 
 
 exports.createSauce = (req, res, next) => {
@@ -22,6 +23,11 @@ exports.createSauce = (req, res, next) => {
   };
 
 exports.modifySauce = (req, res, next) => {
+    const token = req.headers.authorization.split(' ')[1];
+    //Décode le token
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+    //Extrait l'id utilisateur et compare à celui extrait du token
+    const userId = decodedToken.userId;
     //Constante qui regarde si "req.file" existe. Si oui, traite la nouvelle image. Si non, traite l'objet entrant.
     const sauceObject = req.file ?
         {
@@ -29,23 +35,36 @@ exports.modifySauce = (req, res, next) => {
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
         } : { ...req.body };
         //Crée une instance "Sauce" à partir de "sauceObject"
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !'}))
-        .catch(error => res.status(400).json({ error }));
+        if (sauceObject.userId == userId) {
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Objet modifié !'}))
+                .catch(error => res.status(400).json({ error }));
+        } else {
+            res.status(401).json({ message: 'Opération non autorisée !'});
+    }
 };
 
 exports.deleteSauce = (req, res, next) => {
+    const token = req.headers.authorization.split(' ')[1];
+    //Décode le token
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+    //Extrait l'id utilisateur et compare à celui extrait du token
+    const userId = decodedToken.userId;
     //Utilisation de la méthode findOne() du modèle Mongoose qui renvoit la Sauce ayant le même _id que le paramètre de la requête
     Sauce.findOne({ _id: req.params.id })
       .then(sauce => {
-        //Séparation du nom du fichier grâce au "/images/"" contenu dans l'url
-        const filename = sauce.imageUrl.split('/images/')[1];
-        //Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
-        fs.unlink(`images/${filename}`, () => {
-          Sauce.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-            .catch(error => res.status(400).json({ error }));
-        });
+          if (sauce.userId == userId) {
+            //Séparation du nom du fichier grâce au "/images/"" contenu dans l'url
+            const filename = sauce.imageUrl.split('/images/')[1];
+            //Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
+            fs.unlink(`images/${filename}`, () => {
+            Sauce.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+                .catch(error => res.status(400).json({ error }));
+            });
+           } else {
+                res.status(401).json({ message: 'Opération non autorisée !'});
+        }
       })
       .catch(error => res.status(500).json({ error }));
   };
@@ -70,24 +89,42 @@ exports.like = (req, res, next) => {
     //AJOUTER UN LIKE OU UN DISLIKE
 
     if (like === 1) {
-        Sauce.updateOne({_id: req.params.id}, {
-            //Insère le userId dans le tableau usersLiked du modèle
-            $push: {usersLiked: req.body.userId},
-            //Ajoute le likes
-            $inc: {likes: +1},
-        })
-            .then(() => res.status(200).json({message: 'J\'aime !'}))
-            .catch((error) => res.status(400).json({error}));
+        Sauce.findOne({_id: req.params.id})
+            .then((sauce) => {
+                //On regarde si l'utilisateur n'a pas déjà liké ou disliké la sauce
+                if (sauce.usersDisliked.includes(req.body.userId) || sauce.usersLiked.includes(req.body.userId)) {
+                    res.status(401).json({ message: 'Opération non autorisée !'});
+                } else {
+                    Sauce.updateOne({_id: req.params.id}, {
+                        //Insère le userId dans le tableau usersLiked du modèle
+                        $push: {usersLiked: req.body.userId},
+                        //Ajoute le like
+                        $inc: {likes: +1},
+                }) 
+                    .then(() => res.status(200).json({message: 'J\'aime !'}))
+                    .catch((error) => res.status(400).json({error}));
+                }
+            })
+            .catch((error) => res.status(404).json({error}));
     };
     if (like === -1) {
-        Sauce.updateOne({_id: req.params.id}, {
-                //Insère le userId dans le tableau usersDisliked du modèle
-                $push: {usersDisliked: req.body.userId},
-                //Ajoute le dislikes
-                $inc: {dislikes: +1},
-        })
-            .then(() => {res.status(200).json({message: 'Je n\'aime pas !'})})
-            .catch((error) => res.status(400).json({error}));
+        Sauce.findOne({_id: req.params.id})
+            .then((sauce) => {
+                //On regarde si l'utilisateur n'a pas déjà liké ou disliké la sauce
+                if (sauce.usersDisliked.includes(req.body.userId) || sauce.usersLiked.includes(req.body.userId)) {
+                    res.status(401).json({ message: 'Opération non autorisée !'});
+                } else {
+                    Sauce.updateOne({_id: req.params.id}, {
+                        //Insère le userId dans le tableau usersLiked du modèle
+                        $push: {usersDisliked: req.body.userId},
+                        //Ajoute le dislike
+                        $inc: {dislikes: +1},
+                }) 
+                    .then(() => res.status(200).json({message: 'Je n\'aime pas !'}))
+                    .catch((error) => res.status(400).json({error}));
+                }
+            })
+            .catch((error) => res.status(404).json({error}));
     };
 
     //RETIRER SON LIKE OU SON DISLIKE
